SA-8b split PurchasePlanReviewPage
