+++ CHANGELOG.md
@@ -8,6 +8,7 @@ the canonical record.
 
 ## 2026-05 — feedback brief fixes
 
+- **SA-8b / #535** Split PurchasePlanReviewPage to keep sourcing files under 300-LOC headroom cap.
 - **SA-MED / #512** Sourcing cleanup tightened service-layer workspace
   guards, archived-project refresh filtering, raw Decimal wire prices, hashed
   TrustedParts user-agent workspace identifiers, budget-counter locking, and
